[search-ui] add "ask AI assistant" feature
