+++ package.json
@@ -18,7 +18,7 @@
     "@rollup/plugin-typescript": "^12.1.2",
     "@types/react": "^19.1.8",
     "eslint": "^8.57.0",
-    "eslint-config-universe": "^14.0.0",
+    "eslint-config-universe": "^15.0.3",
     "lerna": "^8.2.1",
     "npm-run-all": "^4.1.5",
     "prettier": "^3.5.3",
