+++ packages/search-ui/src/components/BarLoader.tsx
@@ -1,14 +1,15 @@
 import { mergeClasses } from '@expo/styleguide';
 import React from 'react';
 
-type Props = { isLoading?: boolean };
+type Props = { isLoading?: boolean; className?: string };
 
-export const BarLoader = ({ isLoading }: Props) => (
+export const BarLoader = ({ isLoading, className }: Props) => (
   <div
     role="progressbar"
     className={mergeClasses(
       'bg-palette-blue9 h-0.5 absolute mt-[11px] left-px',
-      isLoading && 'animate-[searchUIBarLoader_2s_infinite_ease-in-out]'
+      isLoading && 'animate-[searchUIBarLoader_2s_infinite_ease-in-out]',
+      className
     )}
   />
 );
